Toggle liked tracks in sticky player
