+++ app/src/lib/components/releases/TrackLikeButton.svelte
@@ -1,0 +1,64 @@
+<script lang="ts">
+	import type { SubmitFunction } from '@sveltejs/kit';
+	import type { ReleaseHydrated, TrackRaw } from '../../../../../shared/types';
+	import { enhance } from '$app/forms';
+	import Heart from '../icons/Heart.svelte';
+
+	let {
+		trackID,
+		likedTracks,
+		lightOrDark
+	}: {
+		trackID: string;
+		likedTracks: {
+			track: TrackRaw;
+			release: ReleaseHydrated | null;
+		}[];
+		lightOrDark: 'light' | 'dark';
+	} = $props();
+
+	const handleLikedTrackChange: SubmitFunction = () => {
+		return async ({ update }) => {
+			update();
+		};
+	};
+</script>
+
+<form method="post" action="/likes?/toggleLikedTrack" use:enhance={handleLikedTrackChange}>
+	<input type="hidden" name="trackId" value={trackID} />
+	<button
+		type="submit"
+		name="action"
+		value="toggleLikedTrack"
+		aria-label="Remove from likes tracks"
+		class={lightOrDark}
+	>
+		{#if likedTracks.some((t) => t.track.id === trackID)}
+			<Heart filled />
+		{:else}
+			<Heart />
+		{/if}
+	</button>
+</form>
+
+<style>
+	form {
+		text-align: right;
+	}
+	form {
+		display: flex;
+		justify-content: flex-end;
+	}
+	button {
+		background: none;
+		border: none;
+		cursor: pointer;
+		display: flex;
+	}
+	.light {
+		color: var(--color-text);
+	}
+	.dark {
+		color: #333;
+	}
+</style>
