+++ app/src/lib/components/releases/TracksTableRow.svelte
@@ -3,9 +3,7 @@
 	import { prettifyDuration } from '../../../../../shared/utils';
 	import type { Session } from '@supabase/supabase-js';
 	import ReleaseTrackButton from './ReleaseTrackButton.svelte';
-	import { enhance } from '$app/forms';
-	import type { SubmitFunction } from '@sveltejs/kit';
-	import Heart from '../icons/Heart.svelte';
+	import TrackLikeButton from './TrackLikeButton.svelte';
 
 	const {
 		i = undefined,
@@ -27,12 +25,6 @@
 		}[];
 		showArtist: boolean;
 	} = $props();
-
-	const handleFavTrackChange: SubmitFunction = () => {
-		return async ({ update }) => {
-			update();
-		};
-	};
 </script>
 
 <tr>
@@ -46,40 +38,12 @@
 		<ReleaseTrackButton {track} {release} {profileData} {session} />
 	</td>
 	<td>
-		<form method="post" action="?/toggleLikedTrack" use:enhance={handleFavTrackChange}>
-			<input type="hidden" name="trackId" value={track.id} />
-			<button
-				type="submit"
-				name="action"
-				value="toggleLikedTrack"
-				aria-label="Remove from likes tracks"
-			>
-				{#if likedTracks.some((t) => t.track.id === track.id)}
-					<Heart filled />
-				{:else}
-					<Heart />
-				{/if}
-			</button>
-		</form>
+		<TrackLikeButton trackID={track.id} {likedTracks} lightOrDark={'light'} />
 	</td>
 </tr>
 
 <style>
-	form {
-		text-align: right;
-	}
 	tr:not(:last-child) {
 		border-bottom: 1px solid gray;
-	}
-	form {
-		display: flex;
-		justify-content: flex-end;
-	}
-	button {
-		background: none;
-		border: none;
-		cursor: pointer;
-		color: var(--color-text);
-		display: flex;
 	}
 </style>
