+++ app/src/routes/+layout.svelte
@@ -67,9 +67,10 @@
 	<AudioPlayer
 		userId={data.session?.user.id}
 		userPayPerStream={data.profileData?.pay_per_stream}
-		title={userState.activeSong.title}
+		track={userState.activeSong}
 		release={userState.activeSongRelease}
 		songUrl={userState.activeSongUrl}
+		likedTracks={data.likedTracks}
 	/>
 {/if}
